refactor(sections): extract heading recognition into heading.rs
Pure motion: telling a heading from prose or a wrapped continuation,
splitting a heading that shares its line with a row, the word-grid and
man-page recognizers, and the group label a block's entries carry.
